backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.11.0
django-mysql==4.5.0
djangorestframework==3.13.1
drf-spectacular==0.21.2
drf-yasg==1.20.0
httpie==3.0.2
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
multidict==6.0.2
mysqlclient==2.1.0
packaging==21.3
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0